refactored protocol adapters (sbus, ibus, crsf, sumd)
* moved the interfaces to header files
* introduced option EXTENDED_TRAINER (32 trainer channels, switches for sumd)
* introduced cli interface to get statistics for trainer input
