+++ radio/src/serial.cpp
@@ -314,13 +314,13 @@ static void serialSetupPort(int mode, etx_serial_init& params)
     break;
 
   case UART_MODE_SBUS_TRAINER:
-    params.baudrate = SBUS_BAUDRATE;
+    params.baudrate = SBus::baudrate;
     params.encoding = ETX_Encoding_8E2,
     params.direction = ETX_Dir_RX;
     break;
 
   case UART_MODE_IBUS_TRAINER:
-    params.baudrate = IBUS_BAUDRATE;
+    params.baudrate = IBus::baudrate;
     params.word_length = ETX_WordLength_8;
     params.parity = ETX_Parity_None;
     params.stop_bits = ETX_StopBits_One;
@@ -329,7 +329,7 @@ static void serialSetupPort(int mode, etx_serial_init& params)
     break;
 
   case UART_MODE_CRSF_TRAINER:
-    params.baudrate = CRSF_BAUDRATE;
+    params.baudrate = CRSF::baudrate;
     params.word_length = ETX_WordLength_8;
     params.parity = ETX_Parity_None;
     params.stop_bits = ETX_StopBits_One;
@@ -338,7 +338,7 @@ static void serialSetupPort(int mode, etx_serial_init& params)
     break;
     
   case UART_MODE_SUMD_TRAINER:
-    params.baudrate = SUMD_BAUDRATE;
+    params.baudrate = SumDV3::baudrate;
     params.word_length = ETX_WordLength_8;
     params.parity = ETX_Parity_None;
     params.stop_bits = ETX_StopBits_One;
